+++ edison-mongo/src/main/java/de/otto/edison/mongo/configuration/MongoConfiguration.java
@@ -1,13 +1,18 @@
-package de.otto.edison.mongo;
+package de.otto.edison.mongo.configuration;
 
 import com.mongodb.*;
 import com.mongodb.client.MongoDatabase;
+import de.otto.edison.jobs.repository.JobRepository;
+import de.otto.edison.mongo.jobs.MongoJobRepository;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Repository;
 
 import java.util.Collections;
 import java.util.List;
@@ -93,6 +98,12 @@ public class MongoConfiguration {
     @ConditionalOnMissingBean(MongoDatabase.class)
     public MongoDatabase mongoDatabase() {
         return mongoClient().getDatabase(databaseName);
+    }
+
+    @Bean
+    @ConditionalOnClass(name = "de.otto.edison.jobs.repository.JobRepository")
+    public JobRepository jobRepository(final MongoDatabase mongoDatabase) {
+        return new MongoJobRepository(mongoDatabase);
     }
 
     private List<MongoCredential> getMongoCredentials() {