+++ edison-mongo/src/main/java/de/otto/edison/mongo/jobs/JobStructure.java
@@ -1,4 +1,4 @@
-package de.otto.edison.jobs.repository.mongo;
+package de.otto.edison.mongo.jobs;
 
 enum JobStructure {